[improve] [broker] Let the producer request success at the first time if the previous one is inactive (#21220)
If a producer establishes a new connection when it is reconnecting, while the previous connection is now inactive, the initial request made on the new connection will fail.  This failure will trigger the topic of cleaning up the inactive producers.  However, upon making a second request, the producer will be able to successfully establish a connection and proceed with the operation.

Make the initial request made on the new connection success.

(cherry picked from commit 40f94d5710f66132926f301606a74d24b1098877)